Implement a command to navigate to existing sections
Pressing "g" will ask the user for a RFC document section to navigate
to. The last section that was navigated to is suggested as a default
in the prompt.

* rfc-mode.el (rfc-mode--titles): Variable that keeps a list of
section titles.
(rfc-mode--last-title): Variable that keeps the last section title
that was navigated to.
(rfc-mode-map): Bind command to g.
(rfc-mode-goto-section): New command to navigate to a RFC section.
(rfc-mode-highlight): Generate the rfc-mode--titles info.