Improve doc-strings
The whole first sentence should fit on the first line.  If that makes
the line a bit long then that is unfortunate but better than wrapping
it onto a new line.  When wrapping onto a line then the second line
should never be intended.  When it can be avoided, then long first
lines should be made shorter.