Free rfc-mode-browse of the dependency on Helm
Packages that are not intrinsically tied to some completion mode,
should not depend on such a package, to avoid forcing the authors
preference on innocent users.

This change shouldn't disturb existing users because we keep using
`helm' if it is available.

Usually optional `helm' support should be implemented in a separate
library, which should be distributed as a separate package, but since
here `helm' support consists of only two short functions, we can skip
that here.